Adjust structure to support unit testing in Xcode
