+++ Package.swift
@@ -2,6 +2,9 @@ import PackageDescription
 
 let package = Package(
     name: "VaporApp",
+    targets: [
+        Target(name: "App", dependencies: ["AppLogic"]),
+        ],
     dependencies: [
         .Package(url: "https://github.com/vapor/vapor.git", majorVersion: 1, minor: 1)
     ],
@@ -11,6 +14,5 @@ let package = Package(
         "Localization",
         "Public",
         "Resources",
-        "Tests",
     ]
 )
