+++ Sources/App/main.swift
@@ -1,13 +1,6 @@
 import Vapor
+import AppLogic
 
-let drop = Droplet()
-
-drop.get { req in
-    return try drop.view.make("welcome", [
-    	"message": drop.localization[req.lang, "welcome", "title"]
-    ])
-}
-
-drop.resource("posts", PostController())
+let drop = try configureDroplet()
 
 drop.run()
