+++ Sources/AppLogic/Models/Post.swift
